Rename to ease introduction of RTT functions.
